requirements.txt
aiohttp
discord.py>=1.3.4
requests
nyzostrings